Move new files to correct location and fix imports.
